float32 correction

+++ feudal_trainer.py
@@ -240,7 +240,7 @@ if __name__== "__main__":
     elif args.gym_env == "feudal_spatial":
         upper_level_obs_space = spaces.Box(
             low = -np.inf, high = np.inf, 
-            shape = (24 * (2 + 6),), dtype = np.float64)
+            shape = (24 * (2 + 6),), dtype = np.float32)
         upper_level_action_space = spaces.Box(
             low = -1, high = 1, 
             shape = (48,), dtype = np.float32)
